Search migrated to dbv4
